requirements.txt
# Zabbix Metrics Extractor - Dependencies
# Install with: pip install -r requirements.txt

customtkinter>=5.2.0
pyzabbix>=1.3.0
Pillow>=10.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
reportlab>=4.0.0
matplotlib>=3.7.0

# For building executable (optional)
# pyinstaller>=6.0.0
